added formatting of tooltip for other chart types
